Some update functions are missing but most of them are done. GET POST PUT DELETE working well.

+++ phenolic_api/src/main/java/pt/ipb/phenolic/repos/PhenolicRepository.java
@@ -3,5 +3,5 @@ package pt.ipb.phenolic.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.ipb.phenolic.models.Phenolic;
 
-public interface PhenolicRepository extends JpaRepository<Phenolic, Integer> {
+public interface PhenolicRepository extends JpaRepository<Phenolic, Long> {
 }
