+++ phenolic_api/src/main/java/pt/ipb/phenolic/repos/SourceRepository.java
@@ -3,5 +3,5 @@ package pt.ipb.phenolic.repos;
 import org.springframework.data.jpa.repository.JpaRepository;
 import pt.ipb.phenolic.models.Source;
 
-public interface SourceRepository extends JpaRepository<Source, Integer> {
+public interface SourceRepository extends JpaRepository<Source, Long> {
 }
